+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 - Fix issue with `<SaraChat>` component showing it was constantly loading instead of handling new chat requests
 - Fix icons with the source re-synchronization
+- Remove the `<SaraChat>` component from the individual projects page
 
 ## Version 0.22.0: March 27th, 2024